pyproject.toml
[tool.poetry]
name = "bauman-quest-bot"
version = "0.1.0"
description = ""
authors = ["StanisLove <sta-tun-slav@yandex.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
python-dotenv = "^0.21.1"
typer = "^0.7.0"

[tool.poetry.scripts]
start = "src.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
